pan-sea-backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# Authentication and security
python-jose==3.3.0
PyJWT>=2.10.0
cryptography==45.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary==2.9.9
alembic>=1.13.0
pgvector>=0.2.4

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.0.3
pydantic_core==2.14.1

# AWS S3 dependencies (aligned versions)
boto3==1.39.11
botocore==1.39.11
aioboto3==15.1.0
aiobotocore==2.24.0

# AI/ML dependencies
openai==1.99.6
openai-whisper>=20240930
torch>=2.1.1
torchaudio>=2.1.1
numpy>=1.21.0,<2.0.0
google-generativeai==0.8.3

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0
SpeechRecognition==3.10.0

# File handling
aiofiles==23.2.1
python-multipart==0.0.6

# Document parsing for teacher materials
pypdf==5.1.0
python-pptx==1.0.2
python-docx==1.1.2
PyMuPDF==1.24.10
pdfplumber==0.11.4

# HTTP clients and networking
httpx==0.24.1
requests==2.32.4
websockets==12.0
asgi-lifespan==2.1.0
asgi-correlation-id==4.3.1

# Utility libraries
python-dotenv==1.1.1
python-decouple==3.8
PyYAML==6.0.2
tqdm==4.67.1

# Type hints and validation
typing_extensions==4.14.1
email_validator==2.2.0

# Development and monitoring
uvloop==0.21.0
watchfiles==1.1.0
gunicorn==22.0.0

# Core dependencies
click==8.2.1
certifi==2025.8.3
urllib3==2.5.0
six==1.17.0
packaging==25.0
python-dateutil==2.9.0.post0

# Redis caching
redis==5.0.1
redis[hiredis]==5.0.1

# System monitoring
psutil==5.9.8

# Additional dependencies for specific features
dnspython==2.7.0
cffi==1.17.1
charset-normalizer==3.4.3
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.1
gotrue==1.3.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==0.17.3
httptools==0.6.4
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
postgrest==0.13.2
pyasn1==0.6.1
pycparser==2.22
realtime==1.0.6
rsa==4.9.1
sniffio==1.3.1
storage3==0.6.1
StrEnum==0.4.15
supafunc==0.3.3
anyio==3.7.1
annotated-types==0.7.0